[SUBMIT] Fix a big bug with submit's garbage collector

+++ remote/submitMapCancel.php
@@ -14,17 +14,32 @@
 
     if (isOk($_POST['game_id']))
     {
-        $db->delete('DELETE FROM game WHERE id=' . $_POST['game_id']);
+        $rs = $db->select('SELECT id FROM map WHERE game_id = ' . $_POST['game_id']);
+
+        if ($rs['total'] == 0)
+        {
+            $db->delete('DELETE FROM game WHERE id=' . $_POST['game_id']);
+        }
     }
 
     if (isOk($_POST['gametype_id']))
     {
-        $db->delete('DELETE FROM gametype WHERE id=' . $_POST['gametype_id']);
+        $rs = $db->select('SELECT id FROM map WHERE gametype_id = ' . $_POST['gametype_id']);
+
+        if ($rs['total'] == 0)
+        {
+            $db->delete('DELETE FROM gametype WHERE id=' . $_POST['gametype_id']);
+        }
     }
 
     if (isOk($_POST['author_id']))
     {
-        $db->delete('DELETE FROM author WHERE id=' . $_POST['author_id']);
+        $rs = $db->select('SELECT id FROM map WHERE author_id = ' . $_POST['author_id']);
+
+        if ($rs['total'] == 0)
+        {
+            $db->delete('DELETE FROM author WHERE id=' . $_POST['author_id']);
+        }
     }
 
     if (isOk($_POST['file_id']))
